requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
requests==2.31.0
python-dotenv==1.0.0
bs4==0.0.
azure-cognitiveservices-speech==1.38.0
openai==1.59.3